--- a/clovers/Cargo.toml
+++ b/clovers/Cargo.toml
@@ -19,6 +19,7 @@ traces = ["tracing"]
 [dependencies]
 nalgebra = { version = "0.31.1", features = ["libm"], default-features = false }
 rand = { version = "0.8.5", features = ["small_rng"], default-features = false }
+rand_distr = "0.4.3"
 serde = { version = "1.0.145", features = ["derive"], default-features = false, optional = true }
 # TODO: remove git dependency after there's a new release with https://github.com/hmeyer/stl_io/pull/14
 stl_io = { git = "https://github.com/hmeyer/stl_io", rev = "80cce5351ee4f8eb2c5cfa9114f8c2bf961f62fa", optional = true }