--- a/clovers/src/random.rs
+++ b/clovers/src/random.rs
@@ -1,27 +1,15 @@
 //! Various internal helper functions for getting specific kinds of random values.
 
 use crate::{Float, Vec3, PI};
-use rand::distributions::Uniform;
 use rand::rngs::SmallRng;
 use rand::Rng;
+use rand_distr::{Distribution, UnitBall, UnitDisc};
 
 /// Internal helper. Originally used for lambertian reflection with flaws
 #[must_use]
 #[inline]
 pub fn random_in_unit_sphere(rng: &mut SmallRng) -> Vec3 {
-    let mut position: Vec3;
-    let distribution = Uniform::new(-1.0, 1.0);
-
-    loop {
-        position = Vec3::new(
-            rng.sample(distribution),
-            rng.sample(distribution),
-            rng.sample(distribution),
-        );
-        if position.magnitude_squared() <= 1.0 {
-            return position;
-        }
-    }
+    UnitBall.sample(rng).into()
 }
 
 /// Internal helper. Use this for the more correct "True Lambertian" reflection
@@ -33,19 +21,8 @@ pub fn random_unit_vector(rng: &mut SmallRng) -> Vec3 {
 /// Internal helper.
 #[must_use]
 pub fn random_in_unit_disk(rng: &mut SmallRng) -> Vec3 {
-    let mut position: Vec3;
-    let distribution = Uniform::new(-1.0, 1.0);
-
-    loop {
-        position = Vec3::new(
-            rng.sample(distribution),
-            rng.sample(distribution),
-            0.0, // z component zero
-        );
-        if position.magnitude_squared() <= 1.0 {
-            return position;
-        }
-    }
+    let v: [Float; 2] = UnitDisc.sample(rng);
+    Vec3::new(v[0], v[1], 0.0)
 }
 
 /// Internal helper.